Rework of module nsdbbdb
- properly initializing mutex variables (many thanks to John Buckman)
- support for LMDB (Lightning Memory-Mapped Database)
- Improved configurability for Makefile (providing paths, see examples below)
- modernization of C code (better following style guide)
- use newer API functions.
- Added debug support via Debug(bdb)

BDB has currently an unclear future, since sleepycat stopped
support. Oracle stepped in, but the newer versions use AGPL, which is
less permissive, leading to the decision in Debian to completely phase
out Berkeley DB in 2013.

LMDB is a newer rewrite of the library, which is included in Debian,
used by e.g. OpenLDAP and uses the OpenLDAP Public License, which is
more business-friendly. LMDB is built for read-heavy workloads,
benchmarks show better performance.

The slides in [1] state that in MemcacheDB LMDB reads are 2-200x
faster than BerkeleyDB, writes are 5-900x faster than BerkeleyDB.  The
slide-set contains background, motivation and several benchmarks.

[1] http://www.lmdb.tech/media/20130329-devox-MDB.pdf

 Build with Berkeley DB 6.2.32 (results in nsdbbdb.so)
    make DBINCLUDE=/opt/local/include/db62  DBLIB=/opt/local/lib/db62/

 Build with LMDB (Lightning Memory-Mapped Database) (results in nsdblmdb.so)
    make LMDB=1 DBINCLUDE=/opt/local/include DBLIB=/opt/local/lib/